Reorganizing components and style sheets
