+++ app.json
@@ -4,11 +4,11 @@
     "slug": "WaniKaniPocket",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./src/assets/images/icons/icon.png",
+    "icon": "./assets/images/icons/icon.png",
     "userInterfaceStyle": "light",
     "newArchEnabled": true,
     "splash": {
-      "image": "./src/assets/images/icons/splash-icon.png",
+      "image": "./assets/images/icons/splash-icon.png",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
@@ -17,7 +17,7 @@
     },
     "android": {
       "adaptiveIcon": {
-        "foregroundImage": "./src/assets/images/icons/adaptive-icon.png",
+        "foregroundImage": "./assets/images/icons/adaptive-icon.png",
         "backgroundColor": "#ffffff"
       },
       "edgeToEdgeEnabled": true
