+++ assets/images/icons/adaptive-icon.png
