+++ src/components/home/dashboard/ButtonText.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Text, TextProps } from 'react-native';
-import { HomeButtonStyles } from '../styles/homebutton.styles';
+import { DashboardStyles } from '../../../styles/home/dashboard.styles';
 
 const ButtonText: React.FC<TextProps> = ({ style, children, ...props }) => {
   return (
-    <Text style={[HomeButtonStyles.button_text, style]} {...props}>
+    <Text style={[DashboardStyles.button_text, style]} {...props}>
       {children}
     </Text>
   );
